fix(agent-chat): sidebar resize no longer collapses; time-since in the tree; aligned title bar
- Resizable sidebars stay EXPANDED while dragging the grip (the pointer leaves
  the rail as you widen it) and ignore mouseleave mid-drag, so an unpinned
  sidebar can't collapse out from under the resize; reverts to hover-collapse
  when the drag ends.
- Compact 'time since last interaction' (now / 5m / 3h / 1d / 2w / 3mo / 1y) now
  shows in the ArcSidebar conversation tree too (extracted to lib/relative-time,
  shared with the mission selector); it swaps out for the delete action on hover.
- Conversation title bar is h-12 (was h-10) + solid border so its bottom line
  aligns with the sidebar header dividers across the app width.
- Menu uses 'mission' terminology (Copy mission name/ID) to match the app.
