+++ client/src/locales/en/agent.json
@@ -259,13 +259,13 @@
   },
   "header": {
     "untitled": "Untitled mission",
-    "menu": "Conversation menu",
+    "menu": "Mission menu",
     "rename": "Rename",
-    "renamePlaceholder": "Conversation name",
+    "renamePlaceholder": "Mission name",
     "renameHint": "Double-click to rename",
     "renameFailed": "Could not rename",
-    "copyName": "Copy conversation name",
-    "copyId": "Copy conversation ID",
+    "copyName": "Copy mission name",
+    "copyId": "Copy mission ID",
     "copyProject": "Copy project name",
     "copyProjectPath": "Copy project path",
     "copyFailed": "Couldn't copy"
