+++ client/src/locales/it/agent.json
@@ -259,13 +259,13 @@
   },
   "header": {
     "untitled": "Missione senza titolo",
-    "menu": "Menu della conversazione",
+    "menu": "Menu della missione",
     "rename": "Rinomina",
-    "renamePlaceholder": "Nome della conversazione",
+    "renamePlaceholder": "Nome della missione",
     "renameHint": "Doppio clic per rinominare",
     "renameFailed": "Impossibile rinominare",
-    "copyName": "Copia nome conversazione",
-    "copyId": "Copia ID conversazione",
+    "copyName": "Copia nome missione",
+    "copyId": "Copia ID missione",
     "copyProject": "Copia nome progetto",
     "copyProjectPath": "Copia percorso progetto",
     "copyFailed": "Impossibile copiare"
