+++ client/src/locales/pt/agent.json
@@ -259,13 +259,13 @@
   },
   "header": {
     "untitled": "Missão sem título",
-    "menu": "Menu da conversa",
+    "menu": "Menu da missão",
     "rename": "Renomear",
-    "renamePlaceholder": "Nome da conversa",
+    "renamePlaceholder": "Nome da missão",
     "renameHint": "Duplo clique para renomear",
     "renameFailed": "Não foi possível renomear",
-    "copyName": "Copiar nome da conversa",
-    "copyId": "Copiar ID da conversa",
+    "copyName": "Copiar nome da missão",
+    "copyId": "Copiar ID da missão",
     "copyProject": "Copiar nome do projeto",
     "copyProjectPath": "Copiar caminho do projeto",
     "copyFailed": "Não foi possível copiar"
